+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Open the command output in a new editor
 - Add buttons to toggle history/favorite suggestions
+- Add button to control if command is saved in history
+- Add button to toggle if result is open in a new editor
 
 ### Changed
 - Merge enter command and history mode